Final Version of the Blog Post
Some of the conventions with the divs weren't the cleanest, however, I was able to create a fairly responsive blog post page that was nearly identical to the sketches I drew out prior to coding! I do want to explore more possibilities as I learn more, and refactor the code for practice in the future!