Final Author Bio Page
With this edit, I made the author bio page, which is modestly responsive via media queries, though I had an issue using CSS grid with this page. It kept extending half of the grid beyond the viewport width, so I just used flexbox to organize it to the best of my abilities, though it isn't as well laid out as the blog post page.